# src/api/swagger.json
{"openapi":"3.0.0","paths":{"/habits":{"get":{"operationId":"HabitController_getHabits","parameters":[],"responses":{"200":{"description":"All habits","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/HabitDto"}}}}}}},"post":{"operationId":"HabitController_createHabit","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HabitDto"}}}},"responses":{"200":{"description":"Id of created habit","content":{"application/json":{"schema":{"type":"string"}}}}}}},"/habits/{id}":{"get":{"operationId":"HabitController_getHabitById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Get habit by id","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HabitDto"}}}}}},"put":{"operationId":"HabitController_updateHabit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HabitDto"}}}},"responses":{"200":{"description":"Update habit","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HabitDto"}}}}}},"delete":{"operationId":"HabitController_removeHabit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Delete habit by id","content":{"application/json":{"schema":{"type":"boolean"}}}}}}},"/habits/increment/{id}":{"put":{"operationId":"HabitController_incrementHabit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Increment habit counter","content":{"application/json":{"schema":{"type":"boolean"}}}}}}},"/habits/decrement/{id}":{"put":{"operationId":"HabitController_decrementHabit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Decrement habit counter","content":{"application/json":{"schema":{"type":"boolean"}}}}}}},"/habits/reset/{id}":{"put":{"operationId":"HabitController_resetHabit","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Reset habit counter","content":{"application/json":{"schema":{"type":"boolean"}}}}}}}},"info":{"title":"Habit API","description":"Habit API description","version":"1.0","contact":{}},"tags":[{"name":"habits","description":""}],"servers":[],"components":{"schemas":{"HabitDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"description":{"type":"string"},"icon":{"type":"string"},"currentValue":{"type":"number"},"targetValue":{"type":"number"},"color":{"type":"string"},"period":{"type":"string"}},"required":["id","name","description","icon","currentValue","targetValue","color","period"]}}}}